open/closing logs + linecounter
